requirement.txt
Flask==2.3.2
pytesseract==0.3.10
opencv-python==4.8.0.74
pdf2image==1.16.3
Pillow==10.0.1
numpy==1.26.4
matplotlib==3.8.0
pandas==2.1.2
reportlab==4.0
